requirements.txt
altair==6.0.0
antlr4-python3-runtime==4.9.3
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorlog==6.10.1
flatbuffers==25.12.19
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.18.1
numpy==2.4.4
omegaconf==2.3.0
onnxruntime==1.24.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.1.1
protobuf==6.33.6
pyarrow==23.0.1
pyclipper==1.4.0
pydeck==0.9.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
rapidocr==3.7.0
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
shapely==2.1.2
six==1.17.0
smmap==5.0.3
streamlit==1.55.0
sympy==1.14.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wordninja==2.0.0
